+++ src/domain/folder/handlers.rs
@@ -3,7 +3,9 @@
 //! This module gathers all folder actions triggered by the CLI.
 
 use anyhow::Result;
+use dialoguer::Confirm;
 use himalaya_lib::{AccountConfig, Backend};
+use std::process;
 
 use crate::{
     printer::{PrintTableOpts, Printer},
@@ -11,19 +13,19 @@ use crate::{
 };
 
 pub fn expunge<P: Printer, B: Backend + ?Sized>(
-    folder: &str,
     printer: &mut P,
     backend: &mut B,
+    folder: &str,
 ) -> Result<()> {
     backend.expunge_folder(folder)?;
     printer.print(format!("Folder {folder} successfully expunged!"))
 }
 
 pub fn list<P: Printer, B: Backend + ?Sized>(
-    max_width: Option<usize>,
     config: &AccountConfig,
     printer: &mut P,
     backend: &mut B,
+    max_width: Option<usize>,
 ) -> Result<()> {
     let folders: Folders = backend.list_folders()?.into();
     printer.print_table(
@@ -34,6 +36,33 @@ pub fn list<P: Printer, B: Backend + ?Sized>(
             max_width,
         },
     )
+}
+
+pub fn create<P: Printer, B: Backend + ?Sized>(
+    printer: &mut P,
+    backend: &mut B,
+    folder: &str,
+) -> Result<()> {
+    backend.add_folder(folder)?;
+    printer.print("Folder successfully created!")
+}
+
+pub fn delete<P: Printer, B: Backend + ?Sized>(
+    printer: &mut P,
+    backend: &mut B,
+    folder: &str,
+) -> Result<()> {
+    if let Some(false) | None = Confirm::new()
+        .with_prompt(format!("Confirm deletion of folder {folder}?"))
+        .default(false)
+        .report(false)
+        .interact_opt()?
+    {
+        process::exit(0);
+    };
+
+    backend.delete_folder(folder)?;
+    printer.print("Folder successfully deleted!")
 }
 
 #[cfg(test)]